requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
plotly
newspaper3k
tomli
